Enhance conversational AI: Add Thara persona, AI-powered onboarding parser, and natural language understanding
- Updated bot to introduce itself as 'Thara' throughout
- Created AI-powered onboarding parser for natural language understanding
- Enhanced work hours parsing to handle complex schedules (days, times, travel, classes)
- Improved pillar selection to understand natural language mentions
- Made entire conversation flow more conversational and less robotic
- Added OpenAI-based understanding for all user messages
- Enhanced handle_natural_language to use AI for better understanding
- Updated onboarding messages to be more conversational and friendly
- Fixed onboarding flow to handle natural language responses

All onboarding steps now understand conversational English instead of requiring robotic input formats.

+++ database/models.py
@@ -1,7 +1,7 @@
 """
 SQLAlchemy database models.
 """
-from sqlalchemy import Column, Integer, String, DateTime, Boolean, Text, ForeignKey, Enum as SQLEnum, JSON, Float
+from sqlalchemy import Column, Integer, BigInteger, String, DateTime, Boolean, Text, ForeignKey, Enum as SQLEnum, JSON, Float
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
 from datetime import datetime
@@ -40,7 +40,7 @@ class User(Base):
     __tablename__ = "users"
     
     id = Column(Integer, primary_key=True, index=True)
-    telegram_id = Column(Integer, unique=True, index=True, nullable=False)
+    telegram_id = Column(BigInteger, unique=True, index=True, nullable=False)
     username = Column(String, nullable=True)
     first_name = Column(String, nullable=True)
     last_name = Column(String, nullable=True)
